requirements.txt
sagemaker
flake8
pylint
pytest
PyGithub
pandas
tabulate